+++ src/components/header/header.scss
@@ -1,4 +1,4 @@
-@import "../styles/variables.scss";
+@import "/src/styles/variables.scss";
 
 .header-container {
   background-color: $black;
@@ -21,7 +21,6 @@
     text-indent: -99999px;
     background: #fff;
     .icon-back {
-
     }
   }
   .button-wrapper {
