--- src/components/index.ts
@@ -1,3 +1,0 @@
-import header from "./header";
-
-export { header };
