+++ src/pages/home.ts
@@ -1,4 +1,4 @@
-import { header } from "@/components";
+import header from "@/components/header/header";
 import styles from "./home.scss";
 
 export default function homePage(target: Element) {
